+++ src/router/index.js
@@ -5,9 +5,9 @@ import Login from "../views/Login.vue";
 import Signup from "../views/Signup.vue";
 import MentorInfo from "../views/MentorInfo.vue";
 import LearnerInfo from "../views/LearnerInfo.vue";
-import AdminDashboard from "../views/admin/AdminDashboard.vue";
-import MentorDashboard from "../views/mentor/MentorDashboard.vue";
-import LearnerDashboard from "../views/learner/LearnerDashboard.vue";
+import AdminPage from "../views/admin/AdminPage.vue";
+import MentorPage from "../views/mentor/MentorPage.vue";
+import LearnerPage from "../views/learner/LearnerPage.vue";
 
 const routes = [
   { path: "/", component: Home },
@@ -16,9 +16,9 @@ const routes = [
   { path: "/signup", component: Signup },
   { path: "/mentor-info", component: MentorInfo },
   { path: "/learner-info", component: LearnerInfo },
-  { path: "/admin", component: AdminDashboard },
-  { path: "/mentor", component: MentorDashboard },
-  { path: "/learner", component: LearnerDashboard }
+  { path: "/admin", component: AdminPage },
+  { path: "/mentor", component: MentorPage },
+  { path: "/learner", component: LearnerPage }
 ];
 
 const router = createRouter({
