+++ src/views/Login.vue
@@ -96,7 +96,7 @@ main {
   justify-content: center;
   flex-grow: 1;
   width: 100%;
-  gap: 7rem;
+  gap: 10rem;
   margin-top: 5rem;
 }
 
@@ -119,7 +119,7 @@ main {
   text-align: center;
   max-width: 450px;
   width: 100%;
-  height: 400px;
+  height: 450px;
   background: rgba(6, 102, 120, 0.4);
   padding: 2rem;
   border-radius: 30px;
@@ -131,14 +131,14 @@ main {
   font-size: 1.8rem;
   color: rgb(255, 255, 255);
   margin-bottom: 0.8rem;
-  font-weight: 600;
+  font-weight: bold;
 }
 
 form {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
-  width: 85%;
+  gap: 2rem;
+  width: 95%;
   border-radius: 25px;
   margin-top: 1rem;
 }
@@ -146,16 +146,16 @@ form {
 .input-field {
   display: flex;
   flex-direction: column;
-  gap: 0.3rem;
+  gap: 0.5rem;
   width: 100%;
 }
 
 .input-field label {
   font-size: 0.85rem;
-  font-weight: 500;
+  font-weight: 400;
   text-align: left;
   padding-left: 0.5rem;
-  color:#2d5e7f;
+  color:#ffffff;
 }
 
 .input-with-icon {
@@ -167,7 +167,7 @@ form {
   right: 15px;
   top: 50%;
   transform: translateY(-50%);
-  color: rgba(255, 255, 255, 0.7);
+  color: rgba(165, 161, 161, 0.9);
   cursor: pointer;
 }
 
@@ -177,33 +177,32 @@ input {
   border: 1px solid rgba(255, 255, 255, 0.3);
   border-radius: 25px;
   font-size: 0.95rem;
-  background: rgba(255, 255, 255, 0.2);
-  color: white;
+  background: rgb(255, 255, 255);
   outline: none;
   transition: 0.3s;
   font-family: "Montserrat", sans-serif;
 }
 
 input::placeholder {
-  color: rgba(255, 255, 255, 0.7);
+  color: rgba(62, 81, 90, 0.7);
   font-family: "Montserrat", sans-serif;
   font-size: 13px;
 }
 
 input:focus {
   border-color: rgba(255, 255, 255, 0.6);
-  box-shadow: 0 0 8px rgba(34, 42, 194, 0.5);
+  box-shadow: 0 0 8px rgb(242, 168, 134);
 }
 
 button {
-  background: linear-gradient(to right, #4a98cc, #073544);
-  color: white;
+  background: #ffffff;
+  color: #1c6b79;
   border: none;
   padding: 0.8rem;
   border-radius: 9999px;
   cursor: pointer;
   font-size: 1rem;
-  font-weight: 500;
+  font-weight: bold;
   width: 50%;
   text-align: center;
   display: block;
@@ -214,8 +213,8 @@ button {
 }
 
 button:hover {
-  background: linear-gradient(to right, #073544, #4a98cc);
-  transform: scale(1.05);
+  background: #319cb0;
+  transform: scale(1);
 }
 
  .switch-link {
@@ -225,7 +224,7 @@ button:hover {
 }
 
 .switch-link a {
-  color: #155577;
+  color: #2b617e;
   text-decoration: none;
   font-weight: 500;
 }
