--- src/views/admin/AdminDashboard.vue
@@ -1,12 +1,0 @@
-<template>
-    <div>
-      <h1>This is the ADMIN Dashboard</h1>
-      <p>Only admins should see this page</p>
-    </div>
-  </template>
-  
-  <script>
-  export default {
-    name: 'AdminDashboard'
-  }
-  </script>
